+++ ios/Classes/ImFlutterSdkPlugin.m
@@ -17,7 +17,14 @@
     [registrar publish:wrapper];
 }
 
+@end
 
+@implementation FlutterAppDelegate (Category)
+- (void)applicationDidEnterBackground:(UIApplication *)application {
+    [EMClient.sharedClient applicationDidEnterBackground:application];
+}
 
-
+- (void)applicationWillEnterForeground:(UIApplication *)application {
+    [EMClient.sharedClient applicationWillEnterForeground:application];
+}
 @end